Made the code modular and tried migrating the app due to version issues. Need further More Processing and coding
