+++ android/app/src/main/java/clonewhatsapp/clone_whatsapp/MainActivity.java
@@ -2,6 +2,7 @@ package clonewhatsapp.clone_whatsapp;
 
 import android.os.Bundle;
 import io.flutter.app.FlutterActivity;
+import io.flutter.embedding.android.FlutterActivity;
 import io.flutter.plugins.GeneratedPluginRegistrant;
 
 public class MainActivity extends FlutterActivity {
