+++ lib/models/chatmodel.dart
@@ -1,0 +1,9 @@
+class WhatsAppDetails {
+  String name;
+  String message;
+  String time;
+  MessageStatus status;
+  WhatsAppDetails(this.name, this.message, this.time, this.status);
+}
+
+enum MessageStatus { Delivered, Received, Seen }
